clear bit

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -8,7 +8,7 @@
  * Return: 1 on sucees, -1 on failure
  */
 
-int clear_bit(unsigned int *n, unsigned int index)
+int clear_bit(unsigned long int *n, unsigned int index)
 {
 	unsigned int k;
